some unsuccessful work on the server

+++ jsWebServer/src/shutDown.js
@@ -1,16 +1,21 @@
 const shutdownPack = require("@moebius/http-graceful-shutdown");
 const GracefulShutdownManager = shutdownPack.GracefulShutdownManager;
 
+let shutDownCalled = false;
+
 const shutDown = (server, childArray) => {
-  const shutdownManager = new GracefulShutdownManager(server);
-  shutdownManager.terminate(() => {
-    childArray.forEach((obj) => {
-      console.log(`killing child with id: ${obj.id}`);
-      obj.child.kill();
+  if (!shutDownCalled) {
+    shutDownCalled = true;
+    const shutdownManager = new GracefulShutdownManager(server);
+    shutdownManager.terminate(() => {
+      childArray.forEach((obj) => {
+        console.log(`killing child with id: ${obj.id}`);
+        obj.child.kill();
+      });
+      console.log("shut down complete, exiting...");
+      process.exit(0);
     });
-    console.log("shut down complete, exiting...");
-    process.exit(0);
-  });
+  }
 };
 
 module.exports = { shutDown };
